+++ data-migrations/2024-05-01-1103-create-user-table.sql
@@ -7,6 +7,21 @@ CREATE TABLE `users` (
   `surName` VARCHAR(255) NOT NULL,
   `affiliationId` VARCHAR(255) NOT NULL,
   `acceptedTerms` TINYINT(1) NOT NULL DEFAULT 0,
+  `orcid` VARCHAR(255),
+  `ssoId` VARCHAR(255),
+  `locked` BOOLEAN NOT NULL DEFAULT false,
+  `active` BOOLEAN NOT NULL DEFAULT true,
+
+  `last_sign_in` TIMESTAMP,
+  `last_sign_in_via` VARCHAR(10),
+  `failed_sign_in_attemps` INT NOT NULL DEFAULT 0,
+
+  `notify_on_comment_added` BOOLEAN NOT NULL DEFAULT true,
+  `notify_on_template_shared` BOOLEAN NOT NULL DEFAULT true,
+  `notify_on_feedback_complete` BOOLEAN NOT NULL DEFAULT true,
+  `notify_on_plan_shared` BOOLEAN NOT NULL DEFAULT true,
+  `notify_on_plan_visibility_change` BOOLEAN NOT NULL DEFAULT true,
+
   `created` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
   `createdById` int,
   `modified` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,